Extract recovery_messages.rs from turn.rs (parent #680)
Move the recovery / verifier-repair policy message builders out of
`turn.rs` into a focused sibling module so `turn.rs` keeps shrinking
toward responsibility-focused units (parent meta-issue #680).

Six production helpers + two private sub-builders extracted as free
functions taking `&mut Agent` / `&Agent` (matching `actor_loop_flow` /
`reply_retry` / earlier vertical-slice precedent):

- `focused_edit_no_tool_note_for_target` (target-based focused-edit
  recovery note: missing target vs. compact-read/edit form)
- `focused_edit_no_tool_note_for_policy` (FocusedEditPolicy variant
  with VerifierRepair Read-only / Edit-only / Write-only branches)
- `artifact_directed_recovery_message` (role + allowed tools +
  already-read guidance)
- `verifier_repair_policy_message` (RepairNextAction dispatcher →
  diagnostic / request-patch / safe-stop / transition / setup)
- `verifier_repair_diagnostic_policy_message` (private): diagnostic
  pending note with behavior projection
- `verifier_repair_request_patch_message` (private): existence +
  already-read branches for the patch target
- `artifact_directed_policy_violation_message` (feedback note for a
  policy violation against an artifact-directed target)
- `push_deterministic_ui_recovery_continuation_note` (system-note push
  reminding the model that deterministic UI recovery is recovery, not
  completion)

Behavior unchanged: same prompt text byte-for-byte, same dispatcher
order (RequestDiagnostic / Replan / RequestPatch / SafeStop /
RerunVerifier / VerifiedDone / None+missing_verifier / None default),
same focused_edit_target_already_read consultation, same
focused_edit_policy_violation_feedback_note delegation.

`pub(super)` limited / no facade re-export (DR3-001). Callers updated
in `build_request_messages.rs` (3 sites), `actor_loop_flow.rs` (7
sites) to the free-fn form.

LOC delta:
- turn.rs: 1,839 → 1,646 (-193)
- new module: +225

3,114 lib tests pass; `cargo fmt` / `cargo clippy --lib --all-targets
-- -D warnings` clean.
