+++ src/agent/loop_run/turn.rs
@@ -28,17 +28,13 @@ use super::tool_history::{
     is_preferred_read_edit_target, latest_user_turn_slice, recent_truncated_tool_call_attempt,
 };
 use super::tool_policy::{
-    EffectiveToolPolicy, EffectiveToolPolicyReason, FocusedEditPolicy,
-    effective_tool_policy_error_for_call_with_scope, focused_edit_policy_violation_feedback_note,
+    EffectiveToolPolicy, effective_tool_policy_error_for_call_with_scope,
     workspace_relative_path_for_tool_arg,
 };
 use super::verifier_orchestration::{
     build_verifier_exit_zero_evidence, synthesized_missing_implementation_target_path_for_request,
     task_contract_no_verifier_note, task_contract_verifier_targeted_edit_required_note,
-    verifier_repair_context_diagnostics, verifier_repair_diagnostic_pending_note,
-    verifier_repair_safe_stop_message, verifier_repair_target_display,
-    verifier_repair_transition_message, verifier_repair_unsafe_target_message,
-    verifier_setup_policy_message,
+    verifier_repair_diagnostic_pending_note, verifier_repair_target_display,
 };
 use super::workspace_candidates::existing_workspace_candidate_for_role_in_scope;
 use super::workspace_walk::workspace_appears_empty;
@@ -610,195 +606,6 @@ impl Agent {
             | super::repair_job::RepairNextAction::SafeStop { .. }
             | super::repair_job::RepairNextAction::VerifiedDone => false,
         }
-    }
-
-    pub(super) fn focused_edit_no_tool_note_for_target(
-        &self,
-        target: &Path,
-        target_already_read: bool,
-        attempt: usize,
-    ) -> String {
-        let target_display = progress_path_display(
-            &target.display().to_string(),
-            &self.work_root,
-            self.session.mode_state.active_plan_path.as_deref(),
-            120,
-        );
-        if !target.is_file() {
-            recovery::focused_edit_missing_target_recovery_note(&target_display, attempt)
-        } else {
-            recovery::focused_edit_no_tool_recovery_note(
-                &target_display,
-                target_already_read,
-                attempt,
-            )
-        }
-    }
-
-    pub(super) fn focused_edit_no_tool_note_for_policy(
-        &self,
-        policy: &FocusedEditPolicy,
-        effective_tool_policy: &EffectiveToolPolicy,
-        attempt: usize,
-    ) -> String {
-        let target_display = progress_path_display(
-            &policy.target.display().to_string(),
-            &self.work_root,
-            self.session.mode_state.active_plan_path.as_deref(),
-            120,
-        );
-        if effective_tool_policy.reason() == EffectiveToolPolicyReason::VerifierRepair {
-            match effective_tool_policy.allowed_tool_names_for_prompt() {
-                Some(["Read"]) => {
-                    return format!(
-                        "Verifier repair is waiting for a fresh read of {target_display}. The previous response was not executed. Emit exactly one Read tool call on that file now. Do not call Edit, Bash, Glob, Grep, or answer in prose. verifier_repair_read_attempt={attempt}"
-                    );
-                }
-                Some(["Edit"]) => {
-                    return format!(
-                        "Verifier repair is waiting for a compact edit of {target_display}. The previous response was not executed. Emit exactly one Edit tool call on that file now. Do not call Read, Bash, Glob, Grep, or answer in prose. verifier_repair_edit_attempt={attempt}"
-                    );
-                }
-                Some(["Write"]) => {
-                    return format!(
-                        "Verifier repair is waiting for the missing target {target_display}. The previous response was not executed. Emit exactly one Write tool call on that exact path now. Do not call Read, Bash, Glob, Grep, or answer in prose. verifier_repair_write_attempt={attempt}"
-                    );
-                }
-                _ => {}
-            }
-        }
-        self.focused_edit_no_tool_note_for_target(
-            &policy.target,
-            policy.target_already_read,
-            attempt,
-        )
-    }
-
-    pub(super) fn artifact_directed_recovery_message(
-        &self,
-        effective_tool_policy: &EffectiveToolPolicy,
-    ) -> Option<String> {
-        let policy = effective_tool_policy.artifact_directed_policy()?;
-        let target = self.current_artifact_recovery_target.as_ref()?;
-        let target_display = progress_path_display(
-            &policy.target.display().to_string(),
-            &self.work_root,
-            self.session.mode_state.active_plan_path.as_deref(),
-            120,
-        );
-        let allowed = effective_tool_policy
-            .allowed_tool_names_for_prompt()
-            .map(|tools| tools.join(", "))
-            .unwrap_or_else(|| "Read, Write, Edit".to_string());
-        let read_guidance = if policy.target_already_read {
-            " The target has already been read in this session, so do not call Read again."
-        } else {
-            ""
-        };
-        Some(format!(
-            "[Artifact Directed Recovery] Missing role: {}. Target file: {target_display}. Allowed tools for this turn are {allowed} on that exact target path only.{read_guidance} Do not call Bash, Glob, Grep, or switch files. Use Write if a small scaffold file should be replaced; otherwise use a compact Edit.",
-            target.role.label()
-        ))
-    }
-
-    pub(super) fn verifier_repair_policy_message(
-        &self,
-        effective_tool_policy: &EffectiveToolPolicy,
-    ) -> Option<String> {
-        (effective_tool_policy.reason() == EffectiveToolPolicyReason::VerifierRepair).then(|| {
-            let context = self.repair_job.as_ref();
-            let diagnostics = verifier_repair_context_diagnostics(context);
-            match context.map(|context| context.next_action()) {
-                Some(
-                    super::repair_job::RepairNextAction::RequestDiagnostic
-                    | super::repair_job::RepairNextAction::Replan,
-                ) => self.verifier_repair_diagnostic_policy_message(),
-                Some(super::repair_job::RepairNextAction::RequestPatch { target_hint }) => {
-                    self.verifier_repair_request_patch_message(&diagnostics, &target_hint)
-                }
-                Some(super::repair_job::RepairNextAction::SafeStop { .. }) => {
-                    verifier_repair_safe_stop_message()
-                }
-                Some(
-                    super::repair_job::RepairNextAction::RerunVerifier
-                    | super::repair_job::RepairNextAction::VerifiedDone,
-                ) => verifier_repair_transition_message(),
-                None if self.missing_verifier_job.is_some() => {
-                    verifier_setup_policy_message(&self.active_request_text().unwrap_or_default())
-                }
-                None => verifier_repair_transition_message(),
-            }
-        })
-    }
-
-    fn verifier_repair_diagnostic_policy_message(&self) -> String {
-        let active_request = self.active_request_text().unwrap_or_default();
-        let task_contract = super::task_contract::TaskContract::from_request(&active_request);
-        let behavior_projection =
-            super::required_behavior::project_behavior_contract(&task_contract);
-        self.repair_job
-            .as_ref()
-            .map(|context| {
-                verifier_repair_diagnostic_pending_note(context, behavior_projection.as_ref())
-            })
-            .unwrap_or_else(|| {
-                "[Verifier Repair Policy] A verifier failure is pending. Output a compact diagnosis JSON object only; do not call tools.".to_string()
-            })
-    }
-
-    fn verifier_repair_request_patch_message(
-        &self,
-        diagnostics: &str,
-        target_hint: &super::task_contract::RecoveryTargetHint,
-    ) -> String {
-        let Some(relative) = super::repair_job::safe_relative_path_string(&target_hint.path) else {
-            return verifier_repair_unsafe_target_message();
-        };
-        let target = self.work_root.join(relative);
-        let target = std::fs::canonicalize(&target).unwrap_or(target);
-        let target_display = verifier_repair_target_display(&target, &self.work_root);
-        if !target.is_file() {
-            format!(
-                "[Verifier Repair Policy] A verifier failure is pending.{diagnostics} Missing target file: {target_display}. Next required action: exactly one Write on that target. Do not use Bash, switch files, or finish with prose. Anvil will rerun the verifier after the write."
-            )
-        } else if focused_edit_target_already_read(&self.session.messages, &target, &self.work_root)
-        {
-            format!(
-                "[Verifier Repair Policy] A verifier failure is pending.{diagnostics} Target file: {target_display}. Next required action: exactly one compact Edit on that target. Do not call Read again, Bash, switch files, or finish with prose. Anvil will rerun the verifier after the edit."
-            )
-        } else {
-            format!(
-                "[Verifier Repair Policy] A verifier failure is pending.{diagnostics} Target file: {target_display}. Next required action: exactly one Read on that target. Do not use Edit, Bash, switch files, or finish with prose."
-            )
-        }
-    }
-
-    pub(super) fn artifact_directed_policy_violation_message(
-        &self,
-        effective_tool_policy: &EffectiveToolPolicy,
-    ) -> Option<String> {
-        let policy = effective_tool_policy.artifact_directed_policy()?;
-        let target_display = policy
-            .target
-            .strip_prefix(&self.work_root)
-            .unwrap_or(&policy.target)
-            .to_string_lossy()
-            .replace('\\', "/");
-        focused_edit_policy_violation_feedback_note(
-            &self.session.working_memory.unresolved_errors,
-            effective_tool_policy.allowed_tool_names_for_prompt(),
-            Some(&target_display),
-        )
-    }
-
-    pub(super) fn push_deterministic_ui_recovery_continuation_note(
-        &mut self,
-        target_path: &str,
-        attempt: usize,
-    ) {
-        self.push_system_note(format!(
-            "Deterministic UI recovery updated {target_path}, but this is recovery context, not completion. Inspect the file if needed, then make one small model-produced Edit or run the project verifier before finalizing. deterministic_ui_recovery_attempt={attempt}"
-        ));
     }
 
     pub(super) fn execute_tool_call(
